+++ src/syncer.rs
@@ -1,0 +1,159 @@
+use crate::asyncutil::LimitedShutdownGroup;
+use crate::inventory::InventoryItem;
+use crate::manifest::CsvManifest;
+use crate::s3::S3Client;
+use futures_util::{stream::select, StreamExt};
+use std::fmt;
+use std::fs::File;
+use std::num::NonZeroUsize;
+use std::path::PathBuf;
+use std::sync::Arc;
+use tokio::sync::mpsc::channel;
+use tokio_util::sync::CancellationToken;
+
+#[derive(Debug)]
+pub(crate) struct Syncer {
+    client: Arc<S3Client>,
+    outdir: PathBuf,
+    inventory_jobs: NonZeroUsize,
+    object_jobs: NonZeroUsize,
+}
+
+impl Syncer {
+    pub(crate) fn new(
+        client: S3Client,
+        outdir: PathBuf,
+        inventory_jobs: NonZeroUsize,
+        object_jobs: NonZeroUsize,
+    ) -> Arc<Syncer> {
+        Arc::new(Syncer {
+            client: Arc::new(client),
+            outdir,
+            inventory_jobs,
+            object_jobs,
+        })
+    }
+
+    pub(crate) async fn run(self: &Arc<Self>, manifest: CsvManifest) -> Result<(), MultiError> {
+        let mut inventory_dl_pool = LimitedShutdownGroup::new(self.inventory_jobs.get());
+        let mut object_dl_pool = LimitedShutdownGroup::new(self.object_jobs.get());
+        let (obj_sender, mut obj_receiver) = channel(self.inventory_jobs.get());
+
+        for fspec in manifest.files {
+            let clnt = self.client.clone();
+            let sender = obj_sender.clone();
+            inventory_dl_pool.spawn(move |_| async move {
+                let itemlist = clnt.download_inventory_csv(fspec).await?;
+                for item in itemlist {
+                    let _ = sender.send(item?).await;
+                }
+                Ok(())
+            });
+        }
+
+        let mut errors = Vec::new();
+        let mut inventory_pool_closed = false;
+        let mut object_pool_closed = false;
+        let mut all_objects_txed = false;
+        loop {
+            tokio::select! {
+                r = inventory_dl_pool.next(), if !inventory_pool_closed => {
+                    match r {
+                        Some(Ok(())) => (),
+                        Some(Err(e)) => {
+                            if !errors.is_empty() {
+                                inventory_dl_pool.shutdown();
+                                object_dl_pool.shutdown();
+                            }
+                            errors.push(e);
+                        }
+                        None => inventory_pool_closed = true,
+                    }
+                }
+                r = object_dl_pool.next(), if !object_pool_closed => {
+                    match r {
+                        Some(Ok(())) => (),
+                        Some(Err(e)) => {
+                            if !errors.is_empty() {
+                                inventory_dl_pool.shutdown();
+                                object_dl_pool.shutdown();
+                            }
+                            errors.push(e);
+                        }
+                        None => object_pool_closed = true,
+                    }
+                }
+                r = obj_receiver.recv(), if !all_objects_txed => {
+                    if let Some(item) = r {
+                        let this = self.clone();
+                        object_dl_pool
+                            .spawn(move |token| async move { this.process_item(item, token).await });
+                    } else {
+                        all_objects_txed = true;
+                    }
+                }
+                else => break,
+            }
+        }
+
+        if !errors.is_empty() {
+            Err(MultiError(errors))
+        } else {
+            Ok(())
+        }
+    }
+
+    async fn process_item(
+        self: &Arc<Self>,
+        item: InventoryItem,
+        token: CancellationToken,
+    ) -> anyhow::Result<()> {
+        if token.is_cancelled() {
+            return Ok(());
+        }
+        let url = item.url();
+        let outpath = if item.is_latest {
+            self.outdir.join(&item.key)
+        } else {
+            todo!()
+        };
+        if let Some(p) = outpath.parent() {
+            // TODO: Attach error context:
+            std::fs::create_dir_all(p)?;
+        }
+        // TODO: Attach error context:
+        let outfile = File::create(outpath)?;
+        match token
+            .run_until_cancelled(self.client.download_object(
+                &url,
+                item.details.md5_digest(),
+                &outfile,
+            ))
+            .await
+        {
+            Some(Ok(())) => Ok(()),
+            Some(Err(e)) => Err(e.into()),
+            None => todo!("Delete outfile and empty parent directories"),
+        }
+        // TODO: Manage object metadata and old versions
+        // TODO: Handle concurrent downloads of the same key
+    }
+}
+
+#[derive(Debug)]
+pub(crate) struct MultiError(Vec<anyhow::Error>);
+
+impl fmt::Display for MultiError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let mut first = true;
+        for e in &self.0 {
+            if !std::mem::replace(&mut first, false) {
+                writeln!(f, "\n---")?;
+            }
+            write!(f, "{e:?}")?;
+        }
+        Ok(())
+    }
+}
+
+impl std::error::Error for MultiError {}
